fix: useMount double invocation
Fixes double invocation of useMount effects when used in React 18
strict mode.

See: https://blog.bitsrc.io/react-v18-0-useeffect-bug-why-do-effects-run-twice-39babecede93
